Flatten internal/*
