+++ internal/cli/common.go
@@ -4,7 +4,7 @@ import (
 	"k8s.io/client-go/kubernetes/scheme"
 
 	apiv1 "github.com/substratusai/substratus/api/v1"
-	"github.com/substratusai/substratus/internal/cli/client"
+	"github.com/substratusai/substratus/internal/client"
 )
 
 func init() {
