+++ internal/cli/infer.go
@@ -9,8 +9,8 @@ import (
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
 
-	"github.com/substratusai/substratus/internal/cli/tui"
 	"github.com/substratusai/substratus/internal/cli/utils"
+	"github.com/substratusai/substratus/internal/tui"
 )
 
 func inferCommand() *cobra.Command {
