+++ internal/client/notebook.go
