+++ internal/tui/apply.go
@@ -8,7 +8,7 @@ import (
 	"github.com/charmbracelet/lipgloss"
 	"k8s.io/client-go/kubernetes"
 
-	"github.com/substratusai/substratus/internal/cli/client"
+	"github.com/substratusai/substratus/internal/client"
 )
 
 type ApplyModel struct {
