+++ internal/tui/get.go
@@ -17,7 +17,7 @@ import (
 	"k8s.io/apimachinery/pkg/watch"
 
 	apiv1 "github.com/substratusai/substratus/api/v1"
-	"github.com/substratusai/substratus/internal/cli/client"
+	"github.com/substratusai/substratus/internal/client"
 )
 
 type GetModel struct {
