+++ internal/tui/infer_chat.go
